move dataclasses and errors
